Configure TLS in the Kaniko Chains demo
When using exported OpenShift registry, i.e. on CRC
`default-route-openshift-image-registry.apps-crc.testing`, we need to
configure `kaniko` to trust the ingress CA certificate.

In order to do that the chains-ca-cert secret created in #6 is used and
mounted to the `kaniko-chains` taks using a volume. This diverges from
the default `kaniko-chains` task so we need to patch it with the volume
and volumeMount.

Given that the key holding the trusted certificates in the
`chains-ca-cert` secret is `ca-certificates.crt` this matches one of the
paths that golang would looks at when loading trusted CA certificates
(`/etc/ssl/certs/ca-certificates.crt`), and this makes it nicely
suitable for the mount point.

Ref. https://issues.redhat.com/browse/HACBS-105